Factored out configuration related functionality from the server.
