+++ sys/server/server.cabal
@@ -31,6 +31,7 @@ executable txsserver
   other-modules:        
                         ToProcdef
                       , CmdLineParser
+                      , TxsServerConfig
                       
   build-depends:        base
                       , containers
